blquedeblque
Former-commit-id: c76776f18fab37705ad2c939d48f4b29fcea0612

+++ htdocs/bimpcore/classes/BimpTools.php
@@ -2207,10 +2207,22 @@ class BimpTools
     public static function bloqueDebloque($type, $bloque = true)
     {
         $file = static::getFileBloqued($type);
-        if ($bloque)
-            file_put_contents($file, "Yes");
+        if ($bloque){
+            $random = rand(0,10000000);
+            $text = "Yes".$random;
+            file_put_contents($file, $text);
+            sleep(0.200);
+            $text2 = file_get_contents($file);
+            if($text == $text2)
+                return 1;
+            else{//conflit
+                mailSyn2("Conflit de ref évité", "dev@bimp.fr", "admin@bimp.fr", "Attention : Un conflit de ref de type ".$type." a été évité");
+                self::sleppIfBloqued($type);
+                return static::bloqueDebloque($type, $bloque);
+            }
+        }
         elseif (is_file($file))
-            unlink($file);
+            return unlink($file);
     }
 
     public static function getFileBloqued($type)
@@ -2240,6 +2252,7 @@ class BimpTools
                 dol_syslog("ATTENTION " . $text, 3);
                 mailSyn2("Bloquage anormal", "dev@bimp.fr", "admin@bimp.fr", "Attention : " . $text);
                 static::bloqueDebloque($type, false);
+                return 0;
             }
         } else
             return 0;
